fix: another refactor
