+++ src/components/Combobox/Dynamic/index.ts
@@ -1,0 +1,1 @@
+export * from "./DynamicCombobox";
