+++ src/components/Combobox/Static/Combobox.tsx
@@ -3,17 +3,17 @@ import { forwardRef, InputHTMLAttributes, ReactNode, useRef } from "react";
 
 import { classNames } from "~/utils";
 
-import { Box, List, PropsWithBox, Text } from "..";
-import { HelperText, inputRecipe, InputVariants } from "../BaseInput";
+import { Box, List, PropsWithBox, Text } from "../..";
+import { HelperText, inputRecipe, InputVariants } from "../../BaseInput";
 import {
   listItemStyle,
   listStyle,
   listWrapperRecipe,
   Option,
-} from "../BaseSelect";
+} from "../../BaseSelect";
 
-import { ChangeHandler, useCombobox } from "./useCombobox";
-import { ComboboxWrapper } from "./ComboboxWrapper";
+import { ChangeHandler, useCombobox } from "../useCombobox";
+import { ComboboxWrapper } from "../ComboboxWrapper";
 
 export type ComboboxProps = PropsWithBox<
   Omit<
