+++ src/components/Combobox/Static/index.ts
@@ -1,0 +1,1 @@
+export * from "./Combobox";
