+++ src/components/Multiselect/Dynamic/index.ts
@@ -1,0 +1,1 @@
+export * from "./DynamicMultiselect";
